+++ docs/CHANGELOG.md
@@ -12,7 +12,7 @@
 ## v0.6
 
 * Fixed damage types
-* fixed occasional error with blank form fields causing errors
+* Fixed occasional error with blank form fields causing errors
 
 ## v0.5