+++ docs/CHANGELOG.md
@@ -11,7 +11,8 @@
 
 * Fixed an issue with save DCs being null
 * Added libwrapper as a dependency, and implemented it to cover anything that we wrapped already
-* Updated AC processing to be affected by DAE and display correctly
+* Updated processing for the following items to be affected by DAE and display correctly
+  *  AC, Skills (proficiency and check bonuses), Passive Perception, Saving Throws
 
 ## v0.6